fold disassembly manager into debug session
